refactor(settings): move codelist and licence loading out of settings
Codelists and licences are currently loaded by settings.py and accessed
from there, but this is not best practice.  This commit refactors
settings.py to move these constant pieces of data into constants.py.
We anticipate handling these codelist values separately in the future,
e.g., via a periodic refresh from a separate SSOT and storage in the
database, but for now this cleans up settings.
