+++ iati_account_web/data/views.py
@@ -7,7 +7,7 @@ from django.core.exceptions import SuspiciousOperation
 from django.http import HttpRequest, HttpResponse
 from django.shortcuts import redirect
 from django.template import loader
-from iati_account_web.constants import USER_ROLE_LOOKUP
+from iati_account_web.constants import COUNTRY_LIST, USER_ROLE_LOOKUP
 from iati_account_web.data.forms import (
     CreateDatasetForm,
     CreateOrganisationForm,
@@ -27,7 +27,6 @@ from iati_account_web.ryd_handling.reporting_orgs import (
     parse_dataset_list_to_objects,
     parse_org_list_to_objects,
 )
-from iati_account_web.settings import COUNTRY_LIST
 
 audit_logger = logging.getLogger("audit")
 app_logger = logging.getLogger("iati_account")
