+++ iati_account_web/helpers.py
@@ -1,4 +1,3 @@
-import json
 import logging
 from collections import namedtuple
 
@@ -6,38 +5,6 @@ from django.http import HttpRequest
 from django.shortcuts import redirect
 
 app_logger = logging.getLogger("iati_account")
-
-
-def _codelist_helper(filename: str) -> (list[tuple[str, str]], dict[str, str]):
-    """Helper to load a codelist JSON file and generate a choice list and lookup
-
-    Parameters
-    ----------
-    filename : str
-        JSON codelist filename.
-
-    Returns
-    -------
-    list[tuple[str,str]]
-        Choice list for use in forms.
-    dict[str, str]
-        Lookup mapping codes to names.
-    """
-
-    choice_list = [("", "--")]
-    lookup = {}
-    if filename is not None:
-        with open(filename, "r") as fh:
-            data = json.load(fh)
-
-            choice_list += [(x["code"], x["name"]) for x in data.get("data", [])]
-            choice_list.sort(key=lambda x: x[1])
-
-            lookup = {x["code"]: x["name"] for x in data.get("data", [])}
-
-    lookup[""] = ""
-
-    return choice_list, lookup
 
 
 PreFlightStatus = namedtuple("PreFlightStatus", ["not_okay_to_continue", "okay_to_continue", "redirect"])
